Add prepare_command script to podspec to build hterm if needed

+++ build_hterm.sh
@@ -5,6 +5,12 @@ set -Ceu
 HTERM_VERSION='1.89'
 
 CURRENT_DIR=$(cd "$(dirname "$0")" || exit 1; pwd)
+
+if [ -s "${CURRENT_DIR}/Terminal/Resources/hterm_all.js" ]; 
+	then echo "hterm_all.js exists. skip building else"
+	exit 0
+fi
+
 TMPDIR=$(mktemp -d)
 trap 'rm -rf "${TMPDIR}"' EXIT
